Re-organize modules to avoid cyclic imports
