+++ sdmetrics/datasets/__init__.py
@@ -6,8 +6,9 @@ import os
 from glob import glob
 
 import pandas as pd
+from sdv import Metadata
 
-_dir_ = os.path.dirname(__file__)
+_DIR_ = os.path.dirname(__file__)
 
 
 def list_datasets():
@@ -19,7 +20,7 @@ def list_datasets():
         (List[str]): A list of dataset names.
     """
     datasets = []
-    for path_to_metadata in glob(os.path.join(_dir_, "**/metadata.json")):
+    for path_to_metadata in glob(os.path.join(_DIR_, "**/metadata.json")):
         path_to_dataset = os.path.dirname(path_to_metadata)
         dataset_name = os.path.basename(path_to_dataset)
         datasets.append(dataset_name)
@@ -59,10 +60,7 @@ class Dataset():
         if is_path:
             path_to_dataset = dataset
         else:
-            path_to_dataset = os.path.join(_dir_, dataset)
-
-        # make SDV needed only for testing
-        from sdv import Metadata
+            path_to_dataset = os.path.join(_DIR_, dataset)
 
         metadata = Metadata(os.path.join(path_to_dataset, "metadata.json"))
         tables = Dataset._load_tables(os.path.join(path_to_dataset))
