+++ sdmetrics/detection/tabular/__init__.py
@@ -2,8 +2,8 @@
 This module implements machine learning methods for detecting synthetic
 rows in a single table.
 """
-from .base import TabularDetector
-from .logistic import LogisticDetector
-from .svc import SVCDetector
+from sdmetrics.detection.tabular.base import TabularDetector
+from sdmetrics.detection.tabular.logistic import LogisticDetector
+from sdmetrics.detection.tabular.svc import SVCDetector
 
 __all__ = ["TabularDetector", "LogisticDetector", "SVCDetector"]
