+++ sdmetrics/efficacy/__init__.py
@@ -2,6 +2,6 @@
 This module implements classes/methods to help users evaluate the efficacy
 of the synthetic data (compared to the real data) on various tasks.
 """
-from .base import MLEfficacy
+from sdmetrics.efficacy.base import MLEfficacy
 
 __all__ = ['MLEfficacy']
