+++ sdmetrics/statistical/bivariate/__init__.py
@@ -1,8 +1,9 @@
 """
 This module implements bivariate KL-divergence/relative-entropy measures.
 """
-from .base import BivariateMetric
-from .continuous import ContinuousDivergence
-from .discrete import DiscreteDivergence
+
+from sdmetrics.statistical.bivariate.base import BivariateMetric
+from sdmetrics.statistical.bivariate.continuous import ContinuousDivergence
+from sdmetrics.statistical.bivariate.discrete import DiscreteDivergence
 
 __all__ = ["BivariateMetric", "DiscreteDivergence", "ContinuousDivergence"]
